feat: 페이지 추가(myPage, settingPage)
