--- src/components/Header/types.ts
@@ -1,5 +1,0 @@
-export type HeaderProps = {
-  title: string;
-  leftComponent?: React.ReactElement;
-  rightComponent?: React.ReactElement;
-};
