+++ src/pages/Me/index.tsx
@@ -3,12 +3,12 @@ import { SafeAreaView, Text } from 'react-native';
 import { TextColor } from '~/styles/foundation-color';
 import s from './styles';
 
-const Settings: FC = ({}) => {
+const Me: FC = ({}) => {
   return (
     <SafeAreaView style={s.container}>
-      <Text style={{ color: TextColor['50'] }}>Settings</Text>
+      <Text style={{ color: TextColor['50'] }}>ME</Text>
     </SafeAreaView>
   );
 };
 
-export default Settings;
+export default Me;
