+++ src/pages/Setting/styles.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+const style = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+});
+
+export default style;
